+++ packages/matter.js/src/cluster/export.ts
@@ -4,5 +4,4 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export * from "./server/ClusterServer.js";
-export * from "./server/ClusterServerTypes.js";
+export * from "./server/index.js";