+++ packages/matter.js/src/cluster/server/ClusterServer.ts
@@ -6,14 +6,7 @@
 
 import { capitalize, Diagnostic, ImplementationError, InternalError, Logger, MaybePromise } from "#general";
 import { AccessLevel } from "#model";
-import {
-    ClusterServer as BaseClusterServer,
-    ClusterDatasource,
-    CommandServer,
-    createAttributeServer,
-    createEventServer,
-    Fabric,
-} from "#protocol";
+import { Fabric } from "#protocol";
 import {
     AttributeId,
     BitSchema,
@@ -26,6 +19,8 @@ import {
     TypeFromPartialBitSchema,
 } from "#types";
 import { Endpoint } from "../../device/Endpoint.js";
+import { AnyAttributeServer, createAttributeServer } from "./AttributeServer.js";
+import { ClusterDatasource } from "./ClusterDatasource.js";
 import {
     AttributeInitialValues,
     AttributeServers,
@@ -35,6 +30,8 @@ import {
     EventServers,
     SupportedEventsList,
 } from "./ClusterServerTypes.js";
+import { CommandServer } from "./CommandServer.js";
+import { AnyEventServer, createEventServer } from "./EventServer.js";
 
 const logger = Logger.get("ClusterServer");
 
@@ -48,6 +45,41 @@ function isConditionMatching<F extends BitSchema, SF extends TypeFromPartialBitS
         }
     }
     return false;
+}
+
+/**
+ * A collection of servers for a cluster's attributes, commands and events.
+ */
+export interface BaseClusterServer {
+    /**
+     * Cluster ID
+     */
+    id: ClusterId;
+
+    /**
+     * Cluster name
+     */
+    readonly name: string;
+
+    /**
+     * Cluster datasource
+     */
+    datasource?: ClusterDatasource;
+
+    /**
+     * Cluster attributes as named object that can be used to programmatically work with available attributes
+     */
+    readonly attributes: Record<string, AnyAttributeServer>;
+
+    /**
+     * Cluster commands as array
+     */
+    readonly commands: Record<string, CommandServer>;
+
+    /**
+     * Cluster events as named object
+     */
+    readonly events: Record<string, AnyEventServer>;
 }
 
 /**