+++ packages/node/src/endpoint/server/index.ts
@@ -4,5 +4,4 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export * from "./EndpointServer.js";
 export * from "./ServerEndpointInitializer.js";