+++ CHANGELOG.md
@@ -34,6 +34,7 @@ The main work (all changes without a GitHub username in brackets in the below li
   
 -   @matter/protocol
     - Breaking: `logEndpoint()` was removed. The Endpoints support logging directly via Diagnostics
+    - Breaking: All legacy used *Server classes (ÀttributeServer, EventServer, CommandServer) are moved to the matter.js legacy package
     - Feature: Added `getLocal()` to AttributeClient to retrieve the currently stored/cached value 
     - Adjustment: ACL writes are not sent chunked by default from now on like also in chip SDK
     - Fix: Handles messages only that are secured as required for the relevant protocol